added pr review suggestions
