+++ packages/sketchfab-block/src/SketchfabBlock.tsx
@@ -76,13 +76,14 @@ export const SketchfabBlock = ({ appBridge }: SketchfabBlockProps) => {
                 autospin: bs.autoSpin && bs.autoSpinCount,
                 autostart: bs.autoStart && '1',
                 camera: !bs.startingSpin && '0',
-                dof_circle: (disableAllUI || !bs.uiDOF) && '0',
+                dof_circle: ((bs.accountType === SketchfabAccount.Premium && disableAllUI) || !bs.uiDOF) && '0',
                 fps_speed: bs.fps && bs.fpsValue,
                 max_texture_size: bs.textureSize && bs.textureSizeValue,
                 navigation: bs.navigationMode === SketchfabNavigation.Fps && bs.navigationMode,
                 preload: bs.preloadTextures && '1',
                 scrollwheel: !bs.scrollWheel && '0',
-                ui_stop: (disableAllButtons || !bs.uiDisableViewer) && '0',
+                ui_stop:
+                    ((bs.accountType === SketchfabAccount.Premium && disableAllButtons) || !bs.uiDisableViewer) && '0',
                 ui_theme: bs.uiTheme === SketchfabTheme.Dark && bs.uiTheme,
                 dnt: !bs.viewersTracking && '1',
                 ...annotationSettings,
