+++ packages/sketchfab-block/src/helpers/settings.ts
@@ -1,0 +1,11 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { Bundle } from '@frontify/guideline-blocks-settings';
+import { generateSketchfabEmbedUrl } from './url';
+
+export const parseSketchfabSettingsUrl = (bundle: Bundle) => {
+    const rawUrl = (bundle.getBlock('url')?.value ?? '') as string;
+    const embedUrl = generateSketchfabEmbedUrl(rawUrl);
+
+    bundle.setBlockValue('url', embedUrl);
+};
